+++ mrb_doc/image.rb
@@ -1,14 +1,3 @@
-# Replace the old Colour with the new Colour.
-# @param image [Image]
-# @param old_colour [Colour]
-# @param new_colour [Colour]
-# @return [nil]
-def image_colour_replace!(image, old_colour, new_colour)
-  # mrb_image_colour_replace
-  # src/mruby_integration/image.cpp
-  nil
-end
-
 # Draws the specified portion of the source image into the specified region of
 # the destination image.
 # @param destination [Image]